just eth

+++ com_prop_board.h
@@ -80,12 +80,12 @@ typedef struct __attribute__((__packed__)) {
     uint32_t pid_eth_position_valve_cmd_feed_forward = 0.0;
     uint32_t pid_eth_position_valve_cmd_pid = 0.0;
 
-    uint32_t pid_n2o_mass_flow_measure = 0.0;
-    uint32_t pid_n2o_mass_flow_error = 0.0;
-    uint32_t pid_n2o_mass_flow_error_integration = 0.0;
-    uint32_t pid_n2o_mass_flow_error_derivative = 0.0;
-    uint32_t pid_n2o_position_valve_cmd_feed_forward = 0.0;
-    uint32_t pid_n2o_position_valve_cmd_pid = 0.0;
+    uint16_t pid_n2o_mass_flow_measure = 0.0;
+    uint16_t pid_n2o_mass_flow_error = 0.0;
+    uint16_t pid_n2o_mass_flow_error_integration = 0.0;
+    uint16_t pid_n2o_mass_flow_error_derivative = 0.0;
+    uint16_t pid_n2o_position_valve_cmd_feed_forward = 0.0;
+    uint16_t pid_n2o_position_valve_cmd_pid = 0.0;
 
     
     uint16_t temp_N2O = 0;          // [-70.0;40.0]
